Sheets, required signature boxes on previously filled out forms were always saying that the signature was required even if it had been resigned. Backported to: 14.3.3.0 Reviewed by: Travis
The description of this bug is misleading because I accidentally fixed the above issue in a previous commit when trying to track this one down.  This commit is really for the invalid signature label never showing up for sheets.  Nobody knows about this issue because they couldn't get past the bug mentioned above.